remove unneeded boxing

+++ src/main/java/zowe/client/sdk/zosfiles/uss/input/CreateZfsParams.java
@@ -368,7 +368,7 @@ public class CreateZfsParams {
          */
         public Builder(final Integer cylsPri) {
             ValidateUtils.checkNullParameter(cylsPri == null, "cylsPri is null");
-            ValidateUtils.checkNullParameter(cylsPri.intValue() <= 0, "specify cylsPri greater than 0");
+            ValidateUtils.checkNullParameter(cylsPri <= 0, "specify cylsPri greater than 0");
             this.cylsPri = cylsPri;
         }
